feat : add JSON import/export for students and update grade type
Introduced JsonManager to handle exporting and importing students to/from a JSON file using Jackson. Updated Student and StudentController to use float for grades instead of int, and added a default constructor to Student for Jackson compatibility. Added deleteDatabase method to Database and corresponding test cases for new features. Updated .gitignore and build.gradle for JSON support.

+++ src/main/java/StudentController.java
@@ -12,13 +12,13 @@ public class StudentController {
         this.conn = conn;
     }
 
-    public void createStudent(String firstName, String lastName, int age, int grade) {
+    public void createStudent(String firstName, String lastName, int age, float grade) {
         String sql = "INSERT INTO student (first_name, last_name, age, grade) VALUES (?, ?, ?, ?)";
         try (PreparedStatement pstmt = conn.prepareStatement(sql)) {
             pstmt.setString(1, firstName);
             pstmt.setString(2, lastName);
             pstmt.setInt(3, age);
-            pstmt.setInt(4, grade);
+            pstmt.setFloat(4, grade);
             pstmt.executeUpdate();
             System.out.println("Student created: " + firstName + " " + lastName + ", Age: " + age);
         } catch (SQLException e) {
